+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- GitHub PR comments now use GraphQL to fetch thread resolution and outdated state. Resolved and outdated threads are marked correctly so they can be hidden when **Show resolved/outdated** is set to "hide" (default).
+
 ## [0.17.11] - 2026-02
 
 ### Added